+++ Hankkijogbo/Hankkijogbo/Network/Hankki/HankkiTargetType.swift
@@ -15,8 +15,8 @@ enum HankkiTargetType {
     case getSortOptionFilter
     case getHankkiPin(universityId: Int, storeCategory: String, priceCategory: String, sortOption: String)
     case getHankkiList(universityid: Int, storeCategory: String, priceCategory: String, sortOption: String)
-    case getHankkiThumbnail(id: Int)
-    case getHankkiDetail(id: Int)
+    case getHankkiThumbnail(id: Int64)
+    case getHankkiDetail(id: Int64)
     case postHankkiHeart(id: Int64)
     case deleteHankkiHeart(id: Int64)
     case postHankkiValidate(req: PostHankkiValidateRequestDTO)
