+++ Hankkijogbo/Hankkijogbo/Network/Zip/Request/DeleteZipToHankkRequestiDTO.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct DeleteZipToHankkiRequestDTO: Codable {
     let favoriteId: Int
-    let storeId: Int
+    let storeId: Int64
 }
